requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic[email]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
orjson
huggingface_hub